requirements.txt
openai==1.12.0
streamlit==1.31.0
python-dotenv==1.0.0
pandas==2.2.0
loguru==0.7.2
pydantic==2.5.2
httpx==0.27.0
